Migrate tests

+++ src/pyearth/_types.pyx
@@ -1,5 +1,5 @@
 import numpy as np
 FLOAT = np.float64
-INT = np.int
+INT = int
 INDEX = np.intp
 BOOL = np.uint8
